requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
distro==1.9.0
faiss-cpu==1.12.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==2.3.2
openai==1.100.2
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
